Update app version to 2.1.12
Updated the application version from 1.3.11 to 2.1.12 across multiple files including `ChronoWiz.csproj`, `tizen-manifest.xml`, `Package.appxmanifest`, `app.manifest`, and `MajorMinorBuildRevisionNumbers.xml`. This includes updates to the `ApplicationDisplayVersion`, `ApplicationVersion`, `Version`, and `Identity Version` properties. The build and revision numbers have also been updated from `11` to `12` in the `MajorMinorBuildRevisionNumbers.xml` file. No significant changes have been made to the codebase, as the lines that appear to be removed and added in the `ChronoWiz.csproj` file are identical, suggesting a possible reformatting rather than a change in code.